+++ packages/algob/sample-project/js/test/sample-test.js
@@ -2,6 +2,7 @@ const { getProgram } = require("@algo-builder/runtime");
 const { Runtime, AccountStore } = require("@algo-builder/runtime");
 const { types } = require("@algo-builder/web");
 const { assert } = require("chai");
+const cfg = require("../algob.config");
 
 const minBalance = BigInt(1e6);
 const masterBalance = BigInt(10e6);
@@ -35,6 +36,10 @@ describe("Sample Test", function () {
 		const validTxFee = 10000;
 		assert.equal(fundReceiver.balance(), minBalance);
 		assert.equal(master.balance(), masterBalance);
+
+		// In case you want to get account from algob.config.js with network default
+		// runtime.addAccounts(cfg.networks.default.accounts, 1000000);
+		// console.log(runtime.getAccount(cfg.networks.default.accounts[0].addr));
 
 		runtime.executeTx([
 			{